Login and Owner SIgnUp working

+++ src/main/java/academy/ouaf/dao/BreedDao.java
@@ -3,5 +3,5 @@ package academy.ouaf.dao;
 import academy.ouaf.model.Breed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BreedDao extends JpaRepository<Breed, Long> {
+public interface BreedDao extends JpaRepository<Breed, Short> {
 }
